feat(cnv): conversation module + extract conversation dialogue strings
Create a dedicated cnv (conversation) domain module and pull the conversation
dialogue text that kessel was silently dropping.

The dialogue WAS in the archive all along: 5,768 en-us per-conversation STBs under
/str/cnv/ (the largest /str/ category) -- should_extract_stb only accepted 8
root-level tables + 2 gui ones and rejected every nested table, so str.cnv never
landed in spice. Allow /str/cnv/ now; extract_fqn_from_path already maps the path
onto the cnv.* FQN, so str.cnv.<path>.<id1>.<id2> rejoins its conversation.

New module kessel/src/db/cnv.rs owns the conversation domain (moved out of world.rs,
zero public-API change since impl Database spans files):
- populate_conversation_refs + ConversationRefCounts (the CF-GUID ref graph:
  quest/npc/achievement/codex/item/followup/encounter + alignment-event counts)
- the 8 conversation_* tables
- NEW conversation_lines view: strips str. + trailing .id1.id2 to recover cnv_fqn,
  surfacing dialogue lines per conversation (joinable to conversations -> quests).

world.rs shrinks by 328 lines; cnv populator call site in passes.rs unchanged.

Tests: should_extract_stb accepts /str/cnv/; conversation tables + view exist after
init; conversation_lines recovers cnv_fqn and excludes non-cnv strings.
