+++ kessel/src/db/world.rs
@@ -306,235 +306,6 @@ impl Database {
         Ok(written)
     }
 
-    /// Populate `conversation_quest_refs` by scanning every NODE prototype
-    /// file in `tor_dir` for CF GUID refs that resolve to a known quest.
-    ///
-    /// NODE files at `/resources/systemgenerated/prototypes/<num>.node` hold
-    /// the full conversation playback data for `cnv.*` objects. The PROT
-    /// header (bytes 0x14..) carries the cnv FQN. The body contains CF E0
-    /// GUID refs; those that match a quest GUID indicate the conversation
-    /// grants or otherwise affects that quest. Empirically ~23% of NODE
-    /// files carry such refs.
-    pub fn populate_conversation_refs(
-        &self,
-        tor_dir: &std::path::Path,
-        hashes: &crate::hash::HashDictionary,
-    ) -> Result<ConversationRefCounts> {
-        use crate::myp::Archive;
-        use std::collections::{HashMap, HashSet};
-
-        let conn = self.conn.lock().unwrap();
-
-        // Build a single GUID -> (kind, fqn) map for all objects, so a single
-        // CF E0 scan resolves to its target without per-kind lookups.
-        let guid_to_kind_fqn: HashMap<[u8; 8], (String, String)> = {
-            let mut stmt = conn.prepare("SELECT guid, kind, fqn FROM objects")?;
-            let collected: HashMap<[u8; 8], (String, String)> = stmt
-                .query_map([], |row| {
-                    Ok((
-                        row.get::<_, String>(0)?,
-                        row.get::<_, String>(1)?,
-                        row.get::<_, String>(2)?,
-                    ))
-                })?
-                .filter_map(|r| r.ok())
-                .filter_map(|(guid_hex, kind, fqn)| {
-                    if guid_hex.len() != 16 {
-                        return None;
-                    }
-                    let mut bytes = [0u8; 8];
-                    for i in 0..8 {
-                        bytes[i] = u8::from_str_radix(&guid_hex[i * 2..i * 2 + 2], 16).ok()?;
-                    }
-                    Some((bytes, (kind, fqn)))
-                })
-                .collect();
-            collected
-        };
-
-        let prototype_hashes: HashSet<u64> = hashes
-            .paths_matching("/resources/systemgenerated/prototypes/")
-            .into_iter()
-            .map(|(h, _)| h)
-            .collect();
-
-        let tor_files: Vec<std::path::PathBuf> = std::fs::read_dir(tor_dir)?
-            .filter_map(|e| e.ok())
-            .map(|e| e.path())
-            .filter(|p| p.extension().map(|x| x == "tor").unwrap_or(false))
-            .collect();
-
-        let tx = conn.unchecked_transaction()?;
-        let mut quest_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_quest_refs (cnv_fqn, quest_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut npc_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_npcs (cnv_fqn, npc_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut ach_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_achievements (cnv_fqn, achievement_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut cdx_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_codex (cnv_fqn, codex_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut item_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_items (cnv_fqn, item_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut follow_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_followups (cnv_fqn, target_cnv_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut enc_stmt = tx.prepare_cached(
-            "INSERT OR IGNORE INTO conversation_encounters (cnv_fqn, encounter_fqn) VALUES (?1, ?2)",
-        )?;
-        let mut align_stmt = tx.prepare_cached(
-            "INSERT OR REPLACE INTO conversation_alignment_events (cnv_fqn, event_kind, event_count) VALUES (?1, ?2, ?3)",
-        )?;
-
-        // Alignment-event token kinds. Each entry: (kind_label, byte-needle).
-        // Order matters -- prefix patterns (bigdarkmoment) must come before
-        // their substring patterns (darkmoment) so the more specific bucket
-        // wins.
-        let align_needles: &[(&str, &[u8])] = &[
-            ("bigdarkmoment", b"event.bigdarkmoment"),
-            ("sinistermoment", b"event.sinistermoment"),
-            ("darksidetheme", b"event.darksidetheme"),
-            ("heroicmoment", b"event.heroicmoment"),
-            ("lightsidetheme", b"event.lightsidetheme"),
-            ("darkmoment", b"event.darkmoment"),
-            ("alignment_override", b"alignment_override"),
-            ("influence_desync", b"influence_desync"),
-            ("affection_bot", b"affection_bot"),
-        ];
-
-        let mut counts = ConversationRefCounts::default();
-
-        for tor_path in &tor_files {
-            let mut archive = match Archive::open(tor_path) {
-                Ok(a) => a,
-                Err(_) => continue,
-            };
-            let entries: Vec<_> = match archive.entries() {
-                Ok(e) => e.cloned().collect(),
-                Err(_) => continue,
-            };
-            for entry in &entries {
-                if !prototype_hashes.contains(&entry.filename_hash) {
-                    continue;
-                }
-                let data = match archive.read_entry(entry) {
-                    Ok(d) => d,
-                    Err(_) => continue,
-                };
-
-                let fqn_start = 0x14;
-                if data.len() < fqn_start + 8 {
-                    continue;
-                }
-                let mut fqn_end = fqn_start;
-                while fqn_end < data.len() && fqn_end < fqn_start + 200 && data[fqn_end] != 0 {
-                    fqn_end += 1;
-                }
-                let cnv_fqn = String::from_utf8_lossy(&data[fqn_start..fqn_end]).to_string();
-                if !cnv_fqn.starts_with("cnv.") {
-                    continue;
-                }
-
-                // Per-target dedup: a single conversation often references the
-                // same target multiple times (one per dialog branch); collapse.
-                let mut seen: HashSet<&str> = HashSet::new();
-                let mut i = 0;
-                while i + 9 <= data.len() {
-                    if data[i] == 0xCF && data[i + 1] == 0xE0 {
-                        let mut g = [0u8; 8];
-                        g.copy_from_slice(&data[i + 1..i + 9]);
-                        if let Some((kind, target_fqn)) = guid_to_kind_fqn.get(&g) {
-                            if seen.insert(target_fqn.as_str()) {
-                                match kind.as_str() {
-                                    "Quest" => {
-                                        quest_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.quest += 1;
-                                    }
-                                    "Npc" => {
-                                        npc_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.npc += 1;
-                                    }
-                                    "Achievement" => {
-                                        ach_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.achievement += 1;
-                                    }
-                                    "Codex" => {
-                                        cdx_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.codex += 1;
-                                    }
-                                    "Item" => {
-                                        item_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.item += 1;
-                                    }
-                                    "Conversation" if target_fqn != &cnv_fqn => {
-                                        follow_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.followup += 1;
-                                    }
-                                    "Encounter" => {
-                                        enc_stmt.execute(params![cnv_fqn, target_fqn])?;
-                                        counts.encounter += 1;
-                                    }
-                                    _ => {}
-                                }
-                            }
-                        }
-                        i += 9;
-                    } else {
-                        i += 1;
-                    }
-                }
-
-                // Alignment-event token scan. Walk every printable string in
-                // the NODE, count occurrences per kind, write one row per
-                // (cnv, kind) with the count. The numbered suffixes
-                // (darkmoment_07, heroicmoment_15, ...) collapse into the
-                // unsuffixed kind for storage; downstream can re-scan for
-                // exact tier numbers if needed.
-                let mut align_counts: HashMap<&str, u64> = HashMap::new();
-                let mut si = 0;
-                while si < data.len() {
-                    if (32..127).contains(&data[si]) {
-                        let mut sj = si;
-                        while sj < data.len() && (32..127).contains(&data[sj]) {
-                            sj += 1;
-                        }
-                        if sj - si >= 5 {
-                            let s = &data[si..sj];
-                            for (kind, needle) in align_needles {
-                                if s.windows(needle.len()).any(|w| w == *needle) {
-                                    *align_counts.entry(*kind).or_insert(0) += 1;
-                                    break;
-                                }
-                            }
-                        }
-                        si = sj;
-                    } else {
-                        si += 1;
-                    }
-                }
-                for (kind, n) in &align_counts {
-                    align_stmt.execute(params![cnv_fqn, kind, n])?;
-                    counts.alignment_event += 1;
-                }
-            }
-        }
-
-        drop(quest_stmt);
-        drop(npc_stmt);
-        drop(ach_stmt);
-        drop(cdx_stmt);
-        drop(item_stmt);
-        drop(follow_stmt);
-        drop(enc_stmt);
-        drop(align_stmt);
-        tx.commit()?;
-        Ok(counts)
-    }
-
     /// Populate `quest_chain` with `planet_transition` links by scanning every
     /// `leaving_{planet}` quest for strings that name the destination.
     ///
@@ -659,19 +430,6 @@ impl Database {
     }
 }
 
-/// Per-kind row counts inserted by `populate_conversation_refs`.
-#[derive(Default, Debug)]
-pub struct ConversationRefCounts {
-    pub quest: u64,
-    pub npc: u64,
-    pub achievement: u64,
-    pub codex: u64,
-    pub item: u64,
-    pub followup: u64,
-    pub encounter: u64,
-    pub alignment_event: u64,
-}
-
 /// Convert a `.fxspec` resource path into the path-relative key used by
 /// `<fxSpecString>` references in `.epp` files. Returns None when the
 /// path doesn't contain a `/fxspec/` segment or a `.fxspec` suffix. Used
@@ -723,92 +481,6 @@ pub(crate) fn find_planet_code_after_cc(record: &[u8], cc_id: &[u8; 4]) -> Optio
 pub(crate) fn create_tables(tx: &Transaction) -> Result<()> {
     tx.execute_batch(
         r#"
-            -- Conversation -> quest references. NODE conversation files (cnv.*)
-            -- embed CF GUID refs to qst.* objects representing the quests
-            -- that conversation grants or affects. ~23% of NODE files carry
-            -- such refs in observed data. Populated by scanning .tor archives
-            -- for NODE entries during the populate phase.
-            CREATE TABLE IF NOT EXISTS conversation_quest_refs (
-                cnv_fqn TEXT NOT NULL,
-                quest_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, quest_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_quest_refs_quest ON conversation_quest_refs(quest_fqn);
-            -- Conversation -> NPC actors. CF GUID refs in NODE bodies that
-            -- match npc.* objects. NPC participants in the dialog (the cnv
-            -- FQN's name segment usually picks out the primary NPC; this
-            -- captures every actor present).
-            CREATE TABLE IF NOT EXISTS conversation_npcs (
-                cnv_fqn TEXT NOT NULL,
-                npc_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, npc_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_npcs_npc ON conversation_npcs(npc_fqn);
-            -- Conversation -> achievement unlocks. CF GUID refs to ach.*.
-            CREATE TABLE IF NOT EXISTS conversation_achievements (
-                cnv_fqn TEXT NOT NULL,
-                achievement_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, achievement_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_ach_ach ON conversation_achievements(achievement_fqn);
-            -- Conversation -> codex unlocks. CF GUID refs to cdx.*.
-            CREATE TABLE IF NOT EXISTS conversation_codex (
-                cnv_fqn TEXT NOT NULL,
-                codex_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, codex_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_cdx_cdx ON conversation_codex(codex_fqn);
-            -- Conversation -> item grants. CF GUID refs to itm.* (rewards
-            -- mailed/awarded by the dialog).
-            CREATE TABLE IF NOT EXISTS conversation_items (
-                cnv_fqn TEXT NOT NULL,
-                item_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, item_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_items_item ON conversation_items(item_fqn);
-            -- Conversation -> follow-up conversation. CF GUID refs to other
-            -- cnv.* objects (sequel dialogs, branching outcomes).
-            CREATE TABLE IF NOT EXISTS conversation_followups (
-                cnv_fqn TEXT NOT NULL,
-                target_cnv_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, target_cnv_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_follow_target ON conversation_followups(target_cnv_fqn);
-            -- Conversation -> combat encounter. CF GUID refs to enc.* (combat
-            -- triggered by the dialog).
-            CREATE TABLE IF NOT EXISTS conversation_encounters (
-                cnv_fqn TEXT NOT NULL,
-                encounter_fqn TEXT NOT NULL,
-                PRIMARY KEY (cnv_fqn, encounter_fqn)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_enc_enc ON conversation_encounters(encounter_fqn);
-            -- Per-conversation counts of alignment-event tokens found in NODE
-            -- bytes. SWTOR encodes alignment-coded dialog beats by attaching
-            -- audio/effect event names like `event.darkmoment_NN`,
-            -- `event.bigdarkmoment_NN`, `event.sinistermoment_NN`,
-            -- `event.heroicmoment_NN`, `event.darksidetheme.*`,
-            -- `event.lightsidetheme.*`, plus explicit `alignment_override` and
-            -- `influence_desync` tokens. The presence and count of each kind
-            -- is a coarse signal for the LS/DS/influence character of the
-            -- dialog, even though the per-choice magnitudes (LS+50/+100, etc)
-            -- are not yet decoded.
-            --   event_kind:
-            --     darkmoment        small DS choice trigger
-            --     bigdarkmoment     major DS choice trigger
-            --     sinistermoment    DS choice trigger
-            --     darksidetheme     DS music theme setter
-            --     heroicmoment      LS choice trigger
-            --     lightsidetheme    LS music theme setter
-            --     alignment_override explicit alignment override
-            --     influence_desync  companion influence event
-            --     affection_bot     companion affection-bot reaction
-            CREATE TABLE IF NOT EXISTS conversation_alignment_events (
-                cnv_fqn TEXT NOT NULL,
-                event_kind TEXT NOT NULL,
-                event_count INTEGER NOT NULL,
-                PRIMARY KEY (cnv_fqn, event_kind)
-            );
-            CREATE INDEX IF NOT EXISTS idx_cnv_align_kind ON conversation_alignment_events(event_kind);
             -- Spawn runtime IDs: every SPN triple `spn.X;target.Y;<id>` in a
             -- quest payload becomes one row. The numeric ID may be the runtime
             -- node ID the combat log emits when the entity is interacted with
